// src/store/modules/user/types.ts
export interface UserState {
  name?: string;
  avatar?: string;
  job?: string;
  organization?: string;
  location?: string;
  email?: string;
  introduction?: string;
  personalWebsite?: string;
  jobName?: string;
  organizationName?: string;
  locationName?: string;
  phone?: string;
  registrationDate?: string;
  accountId?: string;
  certification?: number;
  role: Array;
  studentId: string,
  headPortraitUrl: string,
  studentName: string,
  organizationName: string,
  studentNo:string
}